PublisherConfiguration moved to Utilities
