+++ Code/Source/Sensor/SensorConfiguration.h
@@ -7,7 +7,7 @@
  */
 #pragma once
 
-#include "PublisherConfiguration.h"
+#include "Utilities/PublisherConfiguration.h"
 #include <AzCore/RTTI/RTTI.h>
 #include <AzCore/Serialization/SerializeContext.h>
 #include <AzCore/std/containers/map.h>
