+++ Code/Source/Utilities/PublisherConfiguration.h
